+++ src/com/maddyhome/idea/vim/ui/ex/ExActions.kt
@@ -119,7 +119,7 @@ class CompleteEntryAction : TextAction(ExEditorKit.CompleteEntry) {
     //   write action
     // * The key handler routines get the chance to clean up and reset state
     val entry = ExEntryPanel.getInstance().entry
-    KeyHandler.getInstance().handleKey(entry.editor, stroke, entry.context)
+    KeyHandler.getInstance().handleKey(entry.editor, stroke, entry.context, 0)
   }
 
   companion object {