+++ test/org/jetbrains/plugins/ideavim/VimTestCase.kt
@@ -461,7 +461,7 @@ abstract class VimTestCase : UsefulTestCase() {
         val inputModel = TestInputModel.getInstance(editor)
         var key = inputModel.nextKeyStroke()
         while (key != null) {
-          keyHandler.handleKey(editor, key, dataContext)
+          keyHandler.handleKey(editor, key, dataContext, 0)
           key = inputModel.nextKeyStroke()
         }
       }, null, null)